refactor: callback 처리 로직을 hook으로 분리
